requirements.txt
streamlit==1.28.0
langchain==0.0.335
openai==0.28.1
chromadb==0.4.15
pinecone-client==2.2.4
tiktoken==0.5.1
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
plotly==5.17.0
streamlit-chat==0.1.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2